add logic to classes

+++ src/exam/project/IShippingCareStrategy/HaphazardShippingCareStrategy.java
@@ -5,7 +5,7 @@ public class HaphazardShippingCareStrategy implements IShippingCareStrategy {
     private double careCostMultiplier = 0.5;
 
     @Override
-    public double CalculateShippingCare(int shippingMultiplier, double weight, double size) {
+    public double CalculateShippingCare(double shippingMultiplier, double weight, double size) {
         return 0;
     }
