+++ src/exam/project/IShippingCareStrategy/IShippingCareStrategy.java
@@ -1,5 +1,5 @@
 package exam.project.IShippingCareStrategy;
 
 public interface IShippingCareStrategy {
-    double CalculateShippingCare(int shippingMultiplier, double weight, double size);
+    double CalculateShippingCare(double shippingMultiplier, double weight, double size);
 }
