+++ src/exam/project/IShippingTypeStrategy/IShippingTypeStrategy.java
@@ -1,5 +1,5 @@
 package exam.project.IShippingTypeStrategy;
 
 public interface IShippingTypeStrategy {
-    float CalculateShippingMultiplier(int km);
+    double CalculateShippingMultiplier(int km);
 }
